Clean up renaming
