+++ tests/unit/test_netcdf.py
@@ -13,7 +13,7 @@ import pytest
 import xarray as xr
 
 from scmdata import ScmRun
-from scmdata.netcdf import nc_to_run, run_to_nc
+from scmdata.netcdf import _get_xr_dataset_to_write, nc_to_run, run_to_nc
 from scmdata.testing import assert_scmdf_almost_equal
 
 
@@ -667,22 +667,21 @@ def test_run_to_nc_loop_tricky_variable_name(scm_run, start_variable):
     assert_scmdf_almost_equal(scm_run, loaded, check_ts_names=False)
 
 
-def test_run_to_nc_xarray_kwarg_passing(scm_run, tmpdir):
+@patch("scmdata.netcdf._get_xr_dataset_to_write")
+def test_run_to_nc_xarray_kwarg_passing(mock_get_xr_dataset, scm_run, tmpdir):
     dimensions = ["scenario"]
     extras = []
-
     mock_ds = MagicMock()
-    mock_ds.data_vars = scm_run.to_xarray(dimensions, extras).data_vars
-
-    mock_scm_run = MagicMock()
-    mock_scm_run.to_xarray.return_value = mock_ds
+    mock_ds.data_vars = _get_xr_dataset_to_write(scm_run, dimensions, extras).data_vars
+    mock_get_xr_dataset.return_value = mock_ds
 
     out_fname = join(tmpdir, "out.nc")
-    run_to_nc(mock_scm_run, out_fname, dimensions=dimensions, extras=extras, engine="engine")
+    run_to_nc(scm_run, out_fname, dimensions=dimensions, extras=extras, engine="engine")
 
     mock_ds.to_netcdf.assert_called_with(out_fname, engine="engine")
 
 
+@patch("scmdata.netcdf._get_xr_dataset_to_write")
 @pytest.mark.parametrize(
     "in_kwargs,call_kwargs",
     (
@@ -709,19 +708,17 @@ def test_run_to_nc_xarray_kwarg_passing(scm_run, tmpdir):
     ),
 )
 def test_run_to_nc_xarray_kwarg_passing_variable_renaming(
-    scm_run, tmpdir, in_kwargs, call_kwargs
+    mock_get_xr_dataset, scm_run, tmpdir, in_kwargs, call_kwargs
 ):
     dimensions = ["scenario"]
     extras = []
 
     mock_ds = MagicMock()
-    mock_ds.data_vars = scm_run.to_xarray(dimensions, extras).data_vars
-
-    mock_scm_run = MagicMock()
-    mock_scm_run.to_xarray.return_value = mock_ds
+    mock_ds.data_vars = _get_xr_dataset_to_write(scm_run, dimensions, extras).data_vars
+    mock_get_xr_dataset.return_value = mock_ds
 
     out_fname = join(tmpdir, "out.nc")
-    run_to_nc(mock_scm_run, out_fname, dimensions=("scenario",), **in_kwargs)
+    run_to_nc(scm_run, out_fname, dimensions=("scenario",), **in_kwargs)
 
     # variable should be renamed so it matches what goes to disk
     mock_ds.to_netcdf.assert_called_with(out_fname, **call_kwargs)
